refactor(plugin): add PluginApi and PluginHandle, expose on setup hook
