+++ .changes/plugin-setup-refactor.md
@@ -1,0 +1,5 @@
+---
+"tauri": major
+---
+
+Changed the plugin setup hook to take a second argument of type `PluginApi`.
