+++ examples/api/src-tauri/tauri-plugin-sample/src/lib.rs
@@ -1,9 +1,8 @@
 use tauri::{
-  plugin::{Builder, TauriPlugin},
-  Runtime,
+  plugin::{Builder, PluginHandle, TauriPlugin},
+  Manager, Runtime,
 };
 
-const PLUGIN_NAME: &str = "sample";
 #[cfg(target_os = "android")]
 const PLUGIN_IDENTIFIER: &str = "com.plugin.sample";
 
@@ -12,13 +11,20 @@ extern "C" {
   fn init_plugin_sample(webview: tauri::cocoa::base::id);
 }
 
+pub struct SamplePlugin<R: Runtime>(PluginHandle<R>);
+
 pub fn init<R: Runtime>() -> TauriPlugin<R> {
-  Builder::new(PLUGIN_NAME)
-    .setup(|app| {
-      #[cfg(target_os = "android")]
-      app.initialize_android_plugin(PLUGIN_NAME, PLUGIN_IDENTIFIER, "ExamplePlugin")?;
-      #[cfg(target_os = "ios")]
-      app.initialize_ios_plugin(init_plugin_sample)?;
+  Builder::new("sample")
+    .setup(|app, api| {
+      #[cfg(any(target_os = "android", target_os = "ios"))]
+      {
+        #[cfg(target_os = "android")]
+        let handle = api.register_android_plugin(PLUGIN_IDENTIFIER, "ExamplePlugin")?;
+        #[cfg(target_os = "ios")]
+        let handle = api.register_ios_plugin(init_plugin_sample)?;
+        app.manage(SamplePlugin(handle));
+      }
+
       Ok(())
     })
     .build()
