+++ tooling/cli/templates/plugin/src/lib.rs
@@ -5,7 +5,7 @@
 use serde::{ser::Serializer, Serialize};
 use tauri::{
   command,
-  plugin::{Builder, TauriPlugin},
+  plugin::{Builder, PluginHandle, TauriPlugin},
   AppHandle, Manager, Runtime, State, Window,
 };
 
@@ -13,7 +13,6 @@ use std::{collections::HashMap, sync::Mutex};
 
 type Result<T> = std::result::Result<T, Error>;
 
-const PLUGIN_NAME: &str = "{{ plugin_name }}";
 #[cfg(target_os = "android")]
 const PLUGIN_IDENTIFIER: &str = "{{ android_package_id }}";
 
@@ -40,6 +39,8 @@ impl Serialize for Error {
 #[derive(Default)]
 struct MyState(Mutex<HashMap<String, String>>);
 
+pub struct {{ plugin_name_pascal_case }}Plugin<R: Runtime>(PluginHandle<R>);
+
 #[command]
 async fn execute<R: Runtime>(
   _app: AppHandle<R>,
@@ -52,14 +53,18 @@ async fn execute<R: Runtime>(
 
 /// Initializes the plugin.
 pub fn init<R: Runtime>() -> TauriPlugin<R> {
-  Builder::new(PLUGIN_NAME)
+  Builder::new("{{ plugin_name }}")
     .invoke_handler(tauri::generate_handler![execute])
-    .setup(|app| {
+    .setup(|app, api| {
       // initialize mobile plugins
-      #[cfg(target_os = "android")]
-      app.initialize_android_plugin(PLUGIN_NAME, PLUGIN_IDENTIFIER, "ExamplePlugin")?;
-      #[cfg(target_os = "ios")]
-      app.initialize_ios_plugin(init_plugin_{{ plugin_name }})?;
+      #[cfg(any(target_os = "android", target_os = "ios"))]
+      {
+        #[cfg(target_os = "android")]
+        let handle = api.register_android_plugin(PLUGIN_IDENTIFIER, "ExamplePlugin")?;
+        #[cfg(target_os = "ios")]
+        let handle = api.register_ios_plugin(init_plugin_{{ plugin_name }})?;
+        app.manage({{ plugin_name_pascal_case }}Plugin(handle));
+      }
 
       // manage state so it is accessible by the commands
       app.manage(MyState::default());
